Apply patch for VFS-269
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@833575 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs/provider/http/HttpFileObject.java
@@ -120,7 +120,7 @@ public class HttpFileObject
             // Assume 0 content-length
             return 0;
         }
-        return Integer.parseInt(header.getValue());
+        return Long.parseLong(header.getValue());
     }
 
     /**
